lab-controller: Cmd line parsing improvments
Added a new get-serial option that for now does not
do anything.
This commit creates a mutual exclusion group, so
that different actions do not conflict. In this case
activating power is not related to getting information
about serial.

If none of the mandatory options are passed then we abort
with an error.