Added host communication type
The host communication type is useful to
run commands on the target.
It also supports the send expect interface
that the serial communication type does.

One of the functionalities of this commit
is that it enables a device have several stages
for example for the power action.
A dummy host communicaiton type is added to arietta.