Added a separation of io and executable.
Now in config.json definitions for actions
are divided by io and the execute command
to which the io is applicable.

Expect is also a dictionary with information
on the type of match, timeout and the match text.

Serial and usb power were further adapted to use
these new subdivisions.