Distiguish between a stream process and a command
Some commands run for some amount of time and their
success is governed simply by the return status.
On the other hand there are commands which work a bit like
a stream where they block indefenitely waiting for input
or output. An example would be running socat on a serial port.
It is expected that we kill the socat when we are done
and it's exit status is pretty meaningless most if
we killed on purpose. On the other end it is not normal
that a command exits with non 0 status if we do not kill it
and then we should raise the exception.