Removed ran_power checks and throw exception as fast as they come
Previously there was a tolerance for commands to fail
in the power stage. This tolerance is no more and if a
step fails then the whole failure is immediately propagated
to an error.
This cuts a bug where a command exception was caught and
not correctly propagated in the ran_power loops.