do_host_command: removed unused variable and called close before exceptions
The variable killed_on_purpose was not used anywhere and
showed in lint. Therefore it was removed.

On the exception thrown when the application exits with
non zero code we now print the actual exit code as well as
the signal status. We also close the connection just to
be sure.