lab-controller: Add possibility to pass configuration files
In this commit we now allow for command line arguments where the
user passes the configuration json file. To allow for this we need to not print
appliance choices because the choices possible are related to the configuration
file chosen. A small trade-off, but that will allow for more than
one configuration.