config.json: Block for 4 seconds after harware bringup.
This will probably eat a bit of the next application early boot
time, but will fix a crappy problem where the usb network interface
is briefly available, configured and then restarted. This would lead
to an unconfigured interface after this restart and the boot process
would hang as we would not have any network communication with the
arietta.